requirements.txt
Flask==3.0.3
pytest==8.3.2
requests==2.32.3
PyJWT==2.9.0